Expose the compression function for direct chunk write
Implement the jpegls::encodeAsync function to receive a user-provided
raw data (uint8 or uint16) of one chunk in size, decompose it up to 24
subchunks, then compress them to the output buffer.

Enable the function only when c-preprocessor flag `-DH5JPEGLS_USE_ASYNC`
is present in the compile command.

An example app `direct-chunk-write.cpp` is also provided to demonstrate
how compression can be executed asynchronously with the H5Dwrite_chunk
function.